isAging method added

+++ wizard_management/ItchyTamagotchi.java
@@ -2,7 +2,7 @@ package wizard_management;
 
 public class ItchyTamagotchi extends Tamagotchi {
 
-  public ItchyTamagotchi(String name, int hunger_level, int happiness_level, int drama_level, int age){
+  public ItchyTamagotchi(String name, int hunger_level, int happiness_level, int drama_level, double age){
     super(name, hunger_level, happiness_level, drama_level, age);
   }
